Fixed merged conflicts!
