+++ BackEndFlask/core/__init__.py
@@ -9,7 +9,6 @@ app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 app.config['SECRET_KEY'] = 'Thisissupposedtobesecret!'
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///./account.db'
 db = SQLAlchemy(app)
-db.init_app(app)
 ma.ma.init_app(app)
 from controller import bp
 app.register_blueprint(bp, url_prefix='/api')
